backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
SQLAlchemy==2.0.21
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==2.3.7
pytest==7.4.2
pytest-cov==4.1.0
